+++ cmd/cmd.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/sjauld/go-abn/abr"
+	"github.com/aceteknologi/go-abn/abr"
 	"github.com/spf13/cobra"
 )
 
@@ -19,7 +19,7 @@ var rootCmd = &cobra.Command{
 	Use:   "goabn",
 	Short: "Goabn looks up an ABN or ACN using the ABR",
 	Long: `A command line interface to the Australian Business Register.
-            More information available at https://github.com/sjauld/go-abn`,
+            More information available at https://github.com/aceteknologi/go-abn`,
 	RunE: func(cmd *cobra.Command, arg []string) error {
 		fmt.Printf("[DEBUG] %v", arg)
 		return nil