+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## v0.1.0
 
-* Supports the HTTP POST SearchByASICv201408 method including test cases
 * Added support for ACN validation with tests
+* Supports the HTTP POST SearchByASICv201408 method including test cases
+* Added tests for [./abr/abr.go](./abr/abr.go)
 * Mocking of HTTP requests in tests
 
 ## v0.0.1